일정 카드 뷰 수정
